Extract pg-dependent init walproposer code
